win: get rid of some warnings

+++ src/libpmemobj/heap.c
@@ -284,7 +284,7 @@ heap_run_insert(struct bucket *b, uint32_t chunk_id, uint32_t zone_id,
 
 	size_t unit_max = bucket_unit_max(b);
 	struct memory_block m = {chunk_id, zone_id,
-		unit_max - (block_off % 4), block_off};
+		(uint32_t)(unit_max - (block_off % 4)), block_off};
 
 	if (m.size_idx > size_idx)
 		m.size_idx = size_idx;
@@ -293,7 +293,8 @@ heap_run_insert(struct bucket *b, uint32_t chunk_id, uint32_t zone_id,
 		bucket_insert_block(b, m);
 		m.block_off += m.size_idx;
 		size_idx -= m.size_idx;
-		m.size_idx = size_idx > unit_max ? unit_max : size_idx;
+		m.size_idx = size_idx > unit_max ?
+				(uint32_t)unit_max : size_idx;
 	} while (size_idx != 0);
 }
 
@@ -315,7 +316,7 @@ heap_populate_run_bucket(PMEMobjpool *pop, struct bucket *b,
 	ASSERT(hdr->size_idx == 1);
 	ASSERT(bucket_unit_size(b) == run->block_size);
 
-	uint16_t run_bits = RUNSIZE / run->block_size;
+	uint16_t run_bits = (uint16_t)(RUNSIZE / run->block_size);
 	ASSERT(run_bits < (MAX_BITMAP_VALUES * BITS_PER_VALUE));
 	uint16_t block_off = 0;
 	uint16_t block_size_idx = 0;
@@ -576,8 +577,8 @@ heap_drain_to_auxiliary(PMEMobjpool *pop, struct bucket *auxb,
 	int b_id = h->bucket_map[bucket_unit_size(auxb)];
 
 	/* max units drained from a single bucket cache */
-	int units_per_bucket = bucket_bitmap_nallocs(auxb) *
-				MAX_UNITS_PCT_DRAINED_CACHE;
+	int units_per_bucket = (int)(bucket_bitmap_nallocs(auxb) *
+				MAX_UNITS_PCT_DRAINED_CACHE);
 
 	/* max untis drained from all of the bucket caches */
 	int units_total = bucket_bitmap_nallocs(auxb) *
@@ -659,8 +660,8 @@ heap_buckets_init(PMEMobjpool *pop)
 	bucket_proto[i].unit_max = -1;
 	bucket_proto[i].unit_size = CHUNKSIZE;
 
-	h->last_run_max_size = bucket_proto[i - 1].unit_size *
-				(bucket_proto[i - 1].unit_max - 1);
+	h->last_run_max_size = (int)(bucket_proto[i - 1].unit_size *
+				(bucket_proto[i - 1].unit_max - 1));
 
 	h->bucket_map = Malloc(sizeof (*h->bucket_map) * h->last_run_max_size);
 	if (h->bucket_map == NULL)
